+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 v0.3.0 (in development)
 -----------------------
+- Gave `info`, `readme`, and `files` an `--all-versions` option
 
 v0.2.0 (2017-04-03)
 -------------------